Moved StreamSupervisor classes to stream package
