+++ delta/sourcing/src/main/scala/ch/epfl/bluebrain/nexus/sourcing/projections/stream/StatelessStreamSupervisor.scala
@@ -1,10 +1,10 @@
-package ch.epfl.bluebrain.nexus.sourcing.projections
+package ch.epfl.bluebrain.nexus.sourcing.projections.stream
 
 import akka.actor.typed.scaladsl.Behaviors
 import akka.actor.typed.{ActorRef, ActorSystem, SupervisorStrategy}
 import akka.cluster.typed.{ClusterSingleton, SingletonActor}
 import ch.epfl.bluebrain.nexus.delta.kernel.RetryStrategy
-import ch.epfl.bluebrain.nexus.sourcing.projections.StreamSupervisorBehavior._
+import ch.epfl.bluebrain.nexus.sourcing.projections.stream.StreamSupervisorBehavior._
 import fs2.Stream
 import monix.bio.Task
 import monix.execution.Scheduler
