+++ delta/sourcing/src/main/scala/ch/epfl/bluebrain/nexus/sourcing/projections/stream/StreamSupervisor.scala
@@ -1,4 +1,4 @@
-package ch.epfl.bluebrain.nexus.sourcing.projections
+package ch.epfl.bluebrain.nexus.sourcing.projections.stream
 
 import monix.bio.Task
